refactor: integrate link module into link_checker and improve link checking
