+++ rook/src/lib.rs
@@ -3,13 +3,11 @@ pub mod db;
 pub mod domain;
 pub mod email_client;
 pub mod git;
-pub mod link;
-pub mod schedule;
+pub mod link_checker;
 
 pub use configuration::*;
 pub use db::*;
 pub use domain::*;
 pub use email_client::*;
 pub use git::*;
-pub use link::*;
-pub use schedule::*;
+pub use link_checker::*;
